+++ src/routes/Signup/Review/Review.js
@@ -75,7 +75,7 @@ export default class Review extends Component {
   }
 
   componentDidMount = () => {
-    this.props.fetchMySkills()
+    // this.props.fetchMySkills()
   }
 
   getValue = (field) => {
@@ -175,21 +175,6 @@ export default class Review extends Component {
               <span styleName='edit-button text-opacity' onClick={() => this.props.push('/signup/add-location')}>Edit</span>
             </div>
           </div>
-          <div styleName='three-column-input gray-bottom-border'>
-            <div styleName='left-input-column'>
-              <span styleName='text-opacity'>SKILLS</span>
-            </div>
-            <div styleName='center-input-column-scrollable'>
-              {this.props.skills.map((skill, index) =>
-                <div key={index}>
-                  <Pill skill={skill} handlerArg={'name'} />
-                </div>
-              )}
-            </div>
-            <div styleName='right-input-column'>
-              <span styleName='edit-button text-opacity' onClick={this.props.goBack}>Edit</span>
-            </div>
-          </div>
         </div>
         <div>
           <SignupModalFooter submit={this.submit} previous={this.previous} showPrevious={false} continueText={"Let's do this!"} />
@@ -197,10 +182,4 @@ export default class Review extends Component {
       </div>
     </div>
   }
-}
-
-export function Pill ({ skill }) {
-  return <span styleName='review-skill'>
-    {skill.name}
-  </span>
 }